+++ src/expandableAtom/Footer.tsx
